Sketched out a way to implement user-defined queries
Everything will need to be templated on the record type now, but
that's the only way I can allow the user application to specify what
kind of data they want to query and how they will query it. The
application must also provide a "Select Function" that reads data from
a data source, a "Filter Function" (optional) to decide whether a
given data item should be contributed to a query, and an "Aggregate
Function" that determines how to aggregate multiple records from a
query into a final result.