Copied over more of the server, added callback interface for the network
In order to allow the same NetworkManager class to call either
QueryServer or QueryClient's handle_message functions, I needed to
either register each handler individually (so NetworkManager would not
know which class's functions it was calling) or add a common base class
to both of them (so NetworkManager could hold a pointer to the base
class). I chose to add a base class since there are so many handler
functions and they can't be put in a std::map because they all have
slightly different signatures.