Finished implementing crypto, fixed various compile errors
I added new wrapper classes to the openssl module for signing,
verifying, and blind signatures, and also copied in a small C library
that adds blind signatures to OpenSSL itself. Also, started trying to
compile things, and noticed and fixed a lot of compile errors, such as
places where I didn't finish pushing through a change from
std::shared_ptr parameters to reference parameters.

Still not done, becuase AggregationMessageValue and TreeAggregationState
need to be fixed to use the user-supplied AggregateFunction instead of
assuming everything is a vector of FixedPoint.