Fixed TreeAggregation to use DataSource's aggregate function
ProtocolState now needs to hold a reference to the DataSource owned by
QueryClient, since only the DataSource knows how to combine multiple
records into a single aggregate. The DataSource is passed into
TreeAggregationState functions so we can call its aggregate function
when receiving an aggregation message. Also, AggregationMessageValue is
much simpler now, because it just wraps a single value of type
RecordType (which is probably a vector) instead of wrapping a vector.