Applied RecordType template to everything
Once AggregationMessage has the template parameter RecordType, it seems
like there's no way to avoid adding that template parameter to every
other class in the Message and MessageBody hierarchy, even simple things
like ByteBody that ignore the template paramter completely. This is
because the root Message::from_bytes and MessageBody::from_bytes need to
potentially call a subclass from_bytes with a template parameter (e.g.
AggregationMessage<RecordType>::from_bytes), and they have no way of
knowing the template parameter unless it's baked into the Message class.

Now it won't compile because of the circular include dependency between
OverlayMessage and CryptoLibrary. It used to be possible to compile
these separately because OverlayMessage was implemented in a cpp file,
but now the entire implementation of OverlayMessage is in a header that
must be parsed before CryptoLibrary is done being compiled.